DEV-1947 Publish staged event on complete clinical runs
Remove all unused code referencing the archiver command line arguments.

Replace usage of "SbpRestApi" with the new java client in the ClinicalSomaticMetadataApi.

Group shallow seq files into a single event with a new shallow event type.

+++ cluster/src/main/java/com/hartwig/pipeline/metadata/SingleSampleRunMetadata.java
@@ -23,7 +23,7 @@ public interface SingleSampleRunMetadata extends RunMetadata {
 
     @Value.Default
     @JsonIgnore
-    default int entityId() {
+    default long entityId() {
         return -1;
     }
